+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`7492` Added an event direction column (in/out/neutral) to the PnL CSV export.
 * :feature:`11816` Support importing events from Coinledger CSV.
 * :feature:`-` Crosscurve events will now be properly decoded.
 * :feature:`-` Transactions involving stkGHO will now be decoded properly.